requirements.txt
pip == 25.0.1
flask == 2.0.1
werkzeug ==2.0.3
flask-cors == 3.0.10
python-dotenv == 0.19.0
supabase
gunicorn == 21.2.0
scikit-learn
nltk
spacy==3.8.7
stripe
flask-mail==0.9.1
apscheduler==3.10.1